Component iterator fix

+++ app/Main.cpp
@@ -1,25 +1,22 @@
 #include "EntityManager.hpp"
-#include "Test.hpp"
+#include "TestObjects.hpp"
 
 #include <iostream>
 
 int main(int argc, char* argv[]){
 	EntityManager manager;
 	
-	uint32_t id = manager.createEntity();
+	uint64_t id = manager.createEntity();
 	
 	manager.addComponent<Transform>(id, 123.f, 456.f);
 	manager.addComponent<Velocity>(id, 100.f);
+	
+	//manager.setEntityActive(id, false);
 
 	id = manager.createEntity();
 
 	manager.addComponent<Transform>(id, 789.f, 101.f);
 	manager.addComponent<Velocity>(id, 140.f);
-
-	//manager.getComponent<Transform>(id)->x = 0.f;
-	//manager.getComponent<Transform>(id)->y = 0.f;
-
-	//std::cout << "\nDeleting entity " << id << "!\n\n";
 
 	//manager.destroyEntity(id);
